@@ -3,9 +3,14 @@
   "version": "0.1.0",
   "dependencies": {
     "bem-join": "^1.1.2",
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0",
-    "react-scripts-ts": "2.13.0"
+    "lodash": "^4.17.5",
+    "npm": "^5.8.0",
+    "react": "^16.3.1",
+    "react-dom": "^16.3.1",
+    "react-scripts-ts": "^2.13.0"
+  },
+  "peerDependencies": {
+    "lodash": "^4.17"
   },
   "scripts": {
     "start": "react-scripts-ts start",
@@ -14,10 +19,11 @@
     "eject": "react-scripts-ts eject"
   },
   "devDependencies": {
-    "@types/jest": "^22.1.0",
-    "@types/node": "^9.4.0",
-    "@types/react": "^16.0.35",
-    "@types/react-dom": "^16.0.3",
-    "typescript": "^2.6.2"
+    "@types/jest": "^22.2.2",
+    "@types/lodash": "^4.14.106",
+    "@types/node": "^9.6.2",
+    "@types/react": "^16.3.5",
+    "@types/react-dom": "^16.0.4",
+    "typescript": "^2.8.1"
   }
 }
