@@ -1,8 +1,8 @@
 import * as React from 'react'
 import * as ReactDOM from 'react-dom'
-import App from './App'
+import Button from './Button'
 
 it('renders without crashing', () => {
 	const div = document.createElement('div')
-	ReactDOM.render(<App />, div)
+	ReactDOM.render(<Button />, div)
 })
